Added some of the reading and writing functionality
This code is not testsed to see if I can read and or write to a file. I also need to chagne and add of the check to see if data is written to a file and lock them out of submitting another if they already have one. Also need to look into a way of locking them out of the file so they can't write to it.