+ 1 folder google maps and changed user table component
created a new component called map and styles for it, this component parses the longitude and latitude coming in from the database into 2 separate variables, than makes use of the Google maps API and creates a specific map with a marker location corresponding to the longitude and latitude. Changed the user table component to create a popup window with the Google map location whenever Location is clicked.