Robert Dumitrescu final comment
removed update button, instead of it we are calling component did update, to update our data in the tables, also made the tables update whenever acknowledge or solved is clicked. Fixed issues with sign out of admin user, where it would keep the sign up link available on the sign in page. Made the Head ear as a sticky Head-ear so that it is available to the user while scrolling down. Also slightly changed the homepage styling in sign in and sign up to accommodate this change. Fixed issue with routing when signing out from admin while while on Signup page to make it reroute to sign in. Added scroll to top button in the header component to make users life easier, and changed the style of header components. Added google translate in the header component gives user the ability of picking a different language to operate the page with.